Don't use the system NumPy and SciPy packages
The output (`__str__`) of NumPy (and hence SciPy) differs between
versions, in particular in leading whitespace. The system packages for
NumPy and SciPy differ per installed Python 3 version, and as a
result, the doctests cannot be uniformly tested across different
Python versions. Installing NumPy (and SciPy) using pip, obtaining the
a recent version (>= 1.16), bypasses this issue.